+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2020-10-21
+### Added
+- `FirstName`, `LastName`, `Password`, `PasswordHash` and `PasswordSalt` to the user model.
+  Check user model documentation for more details.
+
 ## [1.0.0] - 2020-10-18
 ### Added
 - Initial version